finish1

+++ app/Http/Controllers/UserController.php
@@ -6,7 +6,9 @@ use App\Models\Post;
 use App\Models\Topic;
 use App\Models\User;
 use App\Models\Role;
+use App\Models\News;
 use App\Models\Comment;
+use App\Models\Post_User;
 use App\Models\SubComment;
 use Illuminate\Http\Request;
 use App\Models\Notification;
@@ -18,20 +20,38 @@ class UserController extends Controller
         $user = User::where('id',$id)->first();
         if(Auth::user()->id != $user->id){
             if($rq->radios > Auth::user()->role_id || Auth::user()->role_id <= $user->role_id)
-                return redirect()->route('viewQT');
+                return back();
         }
         else{
             if($rq->radios > Auth::user()->role_id)
-                return redirect()->route('viewQT');
+                return back();
         }
         DB::table('users')->where('id',$id)->update(['role_id'=>$rq->radios]);
         return redirect()->route('viewQT');
     }
     public function deleteUser($id){
         $user = User::where('id',$id)->first();
-        if(Auth::user()->id == $id || Auth::user()->role_id == $user->role_id)
+        if(Auth::user()->id == $id || Auth::user()->role_id == $user->role_id)      
             return redirect()->route('viewQT');
-        DB::table('users')->where('id',$id)->delete();
+
+        $post = Post::where('user_id',$id)->get();
+        foreach($post as $p){
+            $comment = Comment::where('post_id',$p->id)->get();
+            foreach($comment as $c){
+                SubComment::where('comment_id',$c->id)->delete();
+            }
+            Comment::where('post_id',$p->id)->delete();
+        }
+        $comment = Comment::where('user_id',$id)->get();
+        foreach($comment as $c){
+            SubComment::where('comment_id',$c->id)->delete();
+        }
+        Comment::where('user_id',$id)->delete();
+        SubComment::where('user_id',$id)->delete();
+        Post_User::where('user_id',$id)->delete();
+        News::where('user_id',$id)->delete();
+        Post::where('user_id',$id)->delete();
+        User::where('id',$id)->delete();
         return redirect()->route('viewQT');
     }
     public function getinfo($id){
@@ -54,7 +74,7 @@ class UserController extends Controller
     }
     public function updateAvatar(Request $rq, $id){
         $messages=[
-            'fileUpLoad.required'=>'Bạn phải nhập hình tin tức!',
+            'fileUpLoad.required'=>'Bạn phải chọn hình!',
         ];
         $controls = [
             'fileUpLoad'=>'required',
@@ -64,9 +84,11 @@ class UserController extends Controller
         $filename = "logo.png";
         if($rq->file('fileUpLoad')->isValid())
         {
-            if(file_exists(public_path('images/'.$u->avatar)))
-            {
-                unlink(public_path('images/'.$u->avatar));
+            if($u->avatar != 'logo.png'){
+                if(file_exists(public_path('images/'.$u->avatar)))
+                {
+                    unlink(public_path('images/'.$u->avatar));
+                }
             }
             $filename = $rq->fileUpLoad->getClientOriginalName();
             $rq->fileUpLoad->move('images/', $filename);
